% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Cox_Shugart}
\alias{Cox_Shugart}
\title{Cox_Shugart disproportionality index - 1991 ( CS )}
\usage{
Cox_Shugart(votes, seats)
}
\arguments{
\item{votes}{It is a vector of integers, containing the votes of all the candidates.}

\item{seats}{It is a vector of integers, containing the seats obtained by each of the candidates.}
}
\value{
Returns the value obtained for the index
}
\description{
Cox_Shugart's (1991) disproportionality index measures the disproportion in the
distribution of seats by means of a regression line between the percentage of seats and the
percentage of votes. If the distribution were approximately proportional, the slope of the
regression line would be one. If it moves away from that value there will be disproportionality
in the distribution of seats.
}
\examples{

Cox_Shugart(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
