% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Cox_Shugart_correg}
\alias{Cox_Shugart_correg}
\title{Cox_Shugart disproportionality index corrected ( CS_correg )}
\usage{
Cox_Shugart_correg(votes, seats)
}
\arguments{
\item{votes}{It is a vector of integers, containing the votes of all the candidates.}

\item{seats}{It is a vector of integers, containing the seats obtained by each of the candidates.}
}
\value{
Returns the value obtained for the index
}
\description{
This index tries to rectify the defect in the Cox_Shugart index, since the latter is
very sensitive to the presence of small political parties. To correct this deficiency,
this indicator is calculated by not taking into account parties that do not obtain
parliamentary representation. In fact, when its value is calculated the only thing
that is done is to eliminate the votes for parties that do not have obtained any seats and
those results are passed to the "Cox_Shugart()" function.
}
\examples{

Cox_Shugart_correg(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
