% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codigos.R
\name{get_CCAA}
\alias{get_CCAA}
\title{Utilities. ID codes of the Spanish Autonomous Communities}
\usage{
get_CCAA(...)
}
\arguments{
\item{...}{No parameters are required for this function.}
}
\value{
Return a data.frame with two columns. The first column contains the ID code of the
Autonomous Community and the second field contains the name of that region.
}
\description{
With this function you can obtain the ID codes of the Autonomous Communities
of Spain used by the Ministry of Home Affairs to code their data. This codification \href{https://github.com/Miguelro/Electoral/blob/master/Diseno_Registros_MIR/Diseno_Registros.pdf}{can also be seen in this link}.
It should be noted that this coding system does not fully coincide with the
\href{https://www.ine.es/daco/daco42/codmun/cod_ccaa.htm}{Autonomous Community codes of Spanish National Institute of Statistics}.
This function does not take any parameters.
}
\examples{
get_CCAA()



}
