% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codigos.R
\name{get_Provincias}
\alias{get_Provincias}
\title{Utilities. ID codes of the provinces of Spain}
\usage{
get_Provincias(...)
}
\arguments{
\item{...}{No parameters are required for this function.}
}
\value{
It returns a data.frame with two columns. The first column contains the ID code of the
province and the second column contains the name of the province.
}
\description{
With this function, the name of all the Spanish provinces is obtained,
along with their ID codes, which are the same as those provided by the Spanish National
Institute of Statistics \href{https://www.ine.es/daco/daco42/codmun/cod_provincia.htm}{on this website}.
This function does not take any parameters.
}
\examples{
get_Provincias()



}
