% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{IN_LAGO_MONTERO}
\alias{IN_LAGO_MONTERO}
\title{Indice Nacionalizacion de Montero y Lago (2010)}
\usage{
IN_LAGO_MONTERO(Ano, Mes, Ruta, n_escanos = 350)
}
\arguments{
\item{Ano}{Es el valor del año de los datos que se quieren descagar. Este dato puede ser
numérico o de tipo carácter ( mejor numérico)}

\item{Mes}{Es el mes en el que se ha realizado la elección. Obligatoriamente tiene que ser
de tipo carácter y a dos dígitos, es decir "04" es válido pero "4" no es válido}

\item{Ruta}{Se debe indicar una ruta del disco duro local donde se descarga el
fichero del Ministerior del Interior. Una vez finalizada la descarga de forma automática
el fichero descargado se borra.}

\item{n_escanos}{Es el número total de escaños que se deben cubrir. Por defectos tiene
un valor de 350 que son los diputados que se eligen al Congreso de los Diputados en
España.}
}
\value{
El valor devuelto es un objeto de tipo list, con tres posiciones. La primera
posición contiene el valor del índice ( se denomina 'V_indice'), la segunda el vector
conteniendo los tantos por uno de los votos respecto al total nacional
( se denomina Porcentaje_votos'), y el tercer componente son los tantos por uno de escaños
en disputa de las circunscripciones donde se presenta el partido político en
cuestión. Su denominación es 'Porcentaje_escanos'.
}
\description{
Este índice de nacionalización, cuyo valor oscila entre 0 y 1, fue propuesto por
Lago y Montero (2010).Este índice se basa en la decisión de entrada de los partidos en la
competencia electoral en todos los distritos o solo en algunos. Para su cáculo tiene en cuenta
los resultados electorales de los partidos, así como el número de escaños de los distritos
donde se presentan. La fórmula que utiliza es la siguiente:

\deqn{E=\sum_{j=1}^{J}p_{j}^{e}*q_{j}}{E=sum(p_j*q_j;j=1,2,..,J)}

Donde:
p_j es el tanto por uno de los votos obtenidos en todos el territorio sobre el total
q_j es el tanto por uno de escaños ( sobre el total de escaños) de las circunscripciones
en las que la formación política j se presenta.
}
\examples{
s<-IN_LAGO_MONTERO(2019,"04","D:/",n_escanos = 350)
s$V_indice
s$Porcentaje_votos
s$Porcentaje_escanos


}
