% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{L_max}
\alias{L_max}
\title{Index of disproportionality of maximum deviation (Lmax)}
\usage{
L_max(votes, seats)
}
\arguments{
\item{votes}{It is a vector of integers, containing the votes of all the candidates.}

\item{seats}{It is a vector of integers, containing the seats obtained by each of the candidates.}
}
\value{
Returns the value obtained for the index
}
\description{
An indicator of disproportionality that is appropriate for those systems
that use the D'Hondt Law for seats distribution, as is the case in Spain
(except in the Senate elections), presents the disadvantage that it only takes into
account the difference between seats and votes for the most voted political force.
The formula used for its calculation is as follows:
\deqn{Lmax=max[|V_i-E_i]}{Lmax=max[|Vi-Ei|]}
for \eqn{i =  1, 2, \ldots,n }
}
\examples{

L_max(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
