% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{L_Tukey}
\alias{L_Tukey}
\title{Disproportionality index. Tuckey's line (LT)}
\usage{
L_Tukey(votes, seats)
}
\arguments{
\item{votes}{It is a vector of integers, containing the votes of all the candidates.}

\item{seats}{It is a vector of integers, containing the seats obtained by each of the candidates.}
}
\value{
Returns the value obtained for the index
}
\description{
This indicator is also known as the robust bias index. Its value is obtained
by using the slope of the well-known "Tuckey Line" and its use is recommended when there is
a "small" group of parties that behave significantly differently from the rest. This
procedure is qualified as "robust" because the presence of behaviors of certain parties
different from the rest does not substantially alter the line obtained, unlike what happened
when it was used in the least squares criteria to make the regression adjustment.#'
}
\examples{

L_Tukey(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
