% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{nep}
\alias{nep}
\title{Voting dimension index. Effective number of political parties (electoral and parliamentary) (N)}
\usage{
nep(votes, seats)
}
\arguments{
\item{votes}{It is a vector of integers, containing the votes of all the candidates.}

\item{seats}{It is a vector of integers, containing the seats obtained by each of the candidates.}
}
\value{
Returns a dataframe, with the columns "electoral" and "parliamentary" to store the value of
the index of the effective number of parties in the electoral or parliamentary way respectively
}
\description{
This indicator (Laakso and Taagepera 1979) is complementary to the electoral
fragmentation indicator, and its objective is to measure the number of parties that actually
compete in an electoral process. The mathematical formula used to calculate this indicator is as follows:

\deqn{N=\frac{1}{\sum p_{i}^{2}}}{N=1/sum(pi^2)}
for \eqn{i = 1, 2, \ldots,n } and p_i the percentages (as parts per unit) of votes or seats of the party i
}
\examples{

nep(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
