% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{polarizacionDalton}
\alias{polarizacionDalton}
\title{Dimension del voto. Polarizacion ponderada adaptada Dalton. ( Pd )}
\usage{
polarizacionDalton(votosPorcentaje, escala)
}
\arguments{
\item{votosPorcentaje}{Es un vector de números decimales, conteniendo \bold{los porcentajes ( en tantos por ciento ) de votos o escaños} de todas las candidaturas.}

\item{escala}{Es un vector de número decimales entre 1 y 10  para indicar la ubicación ideológica de cada partido
de tal manera que 1 significa totalmente a la izquierda y 10 la ideología es totalmente a
la derecha.}
}
\value{
Devuelve el valor obtenido por el indicador
}
\description{
Con esta función se calcula el índice de polarización ponderada adaptado
de Dalton ( 2008). Este índice pondera las posiciones ideolóicas de los partidos por su
resultado electoral. Este indicador oscila entre 0 ( en el caso hipotético de que todos
los partidos ocupen la misma posición ideológica) y 10 si los partidos se encuentran en los
extremos de la posición de la escala. La fórmula de calculo es la siguiente:

\deqn{Pd=\sqrt{\sum p_{i}\left[\frac{\overline{x}_{i}-\overline{x}_{p}}{4.5}\right]}}{Pd=sqrt(sum(pi*(xi-xp)^2))}
for \eqn{i =1,  2, \ldots,n } y pi los porcentajes  de votos o escaños
del partido i, xi la posición idelógica del partido i  y xp el promedio ideológico
ponderado del sistema de partidos en la escala 1-10
}
\examples{

polarizacionDalton(c(14.12,20.76,10.05,15.42,22.36,3.54),c(6.36,5.31,2.73,4.13,7.04,4.00))

}
