% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leer_Datos.R
\name{Agregado_Mun_MIR}
\alias{Agregado_Mun_MIR}
\title{Download. Datos Agregado a nivel Municipal obtenido del MIR (Ministerior del Interior)}
\usage{
Agregado_Mun_MIR(Ano, Mes, Tipo, Ruta, Borrar = T)
}
\arguments{
\item{Ano}{El año de la eleccion cuatro dígito (YYYY). Puede ser numérico o texto}

\item{Mes}{El mes de la elección. Tiene que se como texto. (Por ejemplo "06", correspondiente a junio)}

\item{Tipo}{El tipo de fichero a descargar: "Congreso" o "Europeas"}

\item{Ruta}{Es la ruta donde se descargarán los ficheros tipo zip del MIR}

\item{Borrar}{Es de tipo Boolean e indica si se borran o no los ficheros después de obtener el data.frame}
}
\value{
Objeto de tipo tbl_df con los datos del voto a nivel municipal
}
\description{
Esta función sirve para descargar los datos en excel a nivel municipal y
carga un data frame con esa información. ¡¡¡Observación!!!: Los campos devueltos son todos
de tipo character, por lo que si es necesario hacer operaciones, habrá que convertir
los campos necsarios a numéricos
}
\examples{
c<-Agregado_Mun_MIR(1989,"06",Tipo = "Europeas","D:/")


}
