% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leer_Datos.R
\name{Agregado_Prov_MIR}
\alias{Agregado_Prov_MIR}
\title{Download. Datos Agregado a nivel Provincial obtenidos del MIR (Ministerio Interior)}
\usage{
Agregado_Prov_MIR(Ano, Mes, Tipo, Ruta, Borrar = T)
}
\arguments{
\item{Ano}{El año de la eleccion cuatro dígito (YYYY). Puede ser numérico o texto}

\item{Mes}{El mes de la elección. Tiene que ser como texto. (Por ejemplo "06", correspondiente a junio)}

\item{Tipo}{El tipo de fichero a descargar: "Congreso" o "Europeas"}

\item{Ruta}{Es la ruta donde se descargarán los ficheros tipo zip del MIR}

\item{Borrar}{Es de tipo Boolean e indica si se borran o no los ficheros después de obtener el data.frame}
}
\value{
Objeto de tipo tbl_df con los datos del voto a nivel Provincial
}
\description{
Esta función sirve para descargar los datos en excel a nivel Provincial y
carga un data frame con esa información.¡¡¡Observación!!!: Los campos devueltos son todos
de tipo character, por lo que si es necesario hacer operaciones, habrá que convertir
los campos necsarios a numéricos
}
\examples{

c<-Agregado_Prov_MIR(2019,"05",Tipo = "Europeas","D:/")


}
