% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arco_parlamentario.R
\name{Arc_Comparacion}
\alias{Arc_Comparacion}
\title{Graficos. Grafico comparativo de Arcos parlamentarios}
\usage{
Arc_Comparacion(Partidos1, Partidos2, Escanos1, Escanos2, cols = NULL,
  repr = c("absolute", "proportion"), titu1 = "Elección1",
  titu2 = "Elección2", titulo = "Poner un titulo")
}
\arguments{
\item{Partidos1}{Es un vector de caracteres con los nombres de los partidos del primer proceso}

\item{Partidos2}{Es un vector de carcateres con los nombres de los partidos del segundo proceso}

\item{Escanos1}{Es un vector de números enteros con los resultados del primer proceso}

\item{Escanos2}{Es un vector de n´meros enteros con los resultados del segundo proceso}

\item{cols}{Es un vector conteniendo los colores a utilizar en la elaboración del gráfico}

\item{repr}{Para indicar "absolute" si se quiere un gráfico en términos absolutos o
"proportion" si se utilizan términos relativos}

\item{titu1}{Expresión a utilizar para nombrar el primer proceso}

\item{titu2}{Expresión a utilizar para nombrar el segundo proceso}

\item{titulo}{Para indicar el título a colocar en el gráfico}
}
\value{
Se obtiene un objeto del tipo ggplot que contiene dos gráficos representativos
de los dos arcos parlamentarios a comparar.
}
\description{
Con este gráfico se obtendrán dos arcos parlamentarios con la finalidad de
poder comparar la estructura de los mismos, es decir los resultados obtenidos en dos
procesos electorales.
}
\examples{

bt1 <- data.frame(parties = c("PP", "CSU", "SPD","IU"),
                 seats   = c(200, 46, 153,58),
                 cols    = c("black", "blue", "red","jjj"),
                 stringsAsFactors = FALSE)

bt2 <- data.frame(parties = c("PP", "CSU","UP1"),
                 seats   = c(100, 146,200),
                 cols    = c("green", "blue","GGG"),
                 stringsAsFactors = FALSE)

Arc_Comparacion(bt1$parties,bt2$parties, bt1$seats,
               bt2$seats, cols=c("PP"="blue", "CSU"="red","SPD"="green",
                                 "IU"="pink","UP1"="antiquewhite" ),
               titu1 = "Abril-2019",titu2 = "Noviembre-2019",
               titu = " Comparación de elecciones")
}
