% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilidades.R
\name{Bazi}
\alias{Bazi}
\title{utilidades.Generacion de ficheros *.bazi}
\usage{
Bazi(Ano, Mes, camino, cota = 3)
}
\arguments{
\item{Ano}{Es el valor del año de los datos que se quieren descagar. Este dato puede ser
numérico o de tipo carácter ( mejor numérico)}

\item{Mes}{Es el mes en el que se ha realizado la elección. Obligatoriamente tiene que ser
de tipo carácter y a dos dígitos, es decir "04" es válido pero "4" no es válido}

\item{cota}{Es la barrera eletoral que se quiera utilizar. En el caso del Conreso de
los Diputados en España es del 3 por ciento (partidos con menos del 3 por ciento de
los votos válidos o entran en el reparto), y es el valor que se le ha asignado por defecto.}

\item{Camino}{Se debe indicar una ruta del disco duro local donde se descarga el
fichero del Ministerior del Interior ( finalizado el proceso se borra de forma automática),
y done se escribirá el fichero con extensión '.bazi' listo para ser usado por BAZI.}
}
\value{
El resultado es un fichero ASCII plano con una estructura de información propia
para ser leido por BAZI y que tiene la siguiente denominación genérica 'Congreso_AAAA_MM.bazi',
donde AAAA es el año del proceso electoral y MM el mes a dos dígitos.
}
\description{
BAZI es un programa Java disponible gratuitamente que implementa varios
métodos de distribución de escaños para sistemas de representación proporcional.
Ofrece la posibilidad de utilizar método de los divisores, así como métodos de cuota.
La descarga del programa se puede hacer en \code{\url{https://www.math.uni-augsburg.de/htdocs/emeriti/pukelsheim/bazi/}}
Instrucciones sobre su uso se pueden en contrar en la siguiente dirección web:
\code{\url{https://opus.bibliothek.uni-augsburg.de/opus4/frontdoor/index/index/docId/601}}.
Con esta función lo que se obtienen son fichero de tipo ASCII planos, con extensión '.bazi'
que tienen una estructura enfocada a que su información pueda ser entendida por
BAZI y obtener los resultados que se pueden entresacar con esta aplicación.
}
\examples{
Bazi(Ano = 2019,"04","D:/")

}
