% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilidades.R
\name{Bochsler}
\alias{Bochsler}
\title{utilidades.Obtener datos para metodo Bochsler (2010)}
\usage{
Bochsler(Ano, Mes, RutaDescarga, RutaSalida)
}
\arguments{
\item{Ano}{Es el valor del año de los datos que se quieren descagar. Este dato puede ser
numérico o de tipo carácter ( mejor numérico)}

\item{Mes}{Es el mes en el que se ha realizado la elección. Obligatoriamente tiene que ser
de tipo carácter y a dos dígitos, es decir "04" es válido pero "4" no es válido}

\item{RutaDescarga}{Se debe indicar una ruta del disco duro local donde se descarga el
fichero del Ministerior del Interior. Una vez finalizada la descarga de forma automática
el fichero descargado se borra.}

\item{RutaSalida}{Es una ruta del disco duro local donde se depositará la hoja excel generada
con los datos para llegar a la macro de Daniel Bochsler. El libro excel generado se
denomina  'Bochsler.xlsx'}
}
\value{
Con la presente función se obtiene una hoja excel denominada 'Bochsler.xlsx' colocada
en la ruta que se ha indicado con el parámetro 'RutaSalida'
}
\description{
El índice de nacionalización de partidos estandarizado elaborado por Bochsler (2010)
se puede calcular mediante una hoja de cálculo excel, que se puede descargar de la siguiente
página web \code{\url{https://www.bochsler.eu/pns/}}. El fichero excel que se puede desccargar
contiene una macro para clacular \bold{el índice de nacionalización de partidos}, muy
utilizado en la actualidad. El objetivo de esta función es crear una hoja excel que se
carga directamente con los datos facilitados por el Ministerior del Interior (MIR) de españa,
\code{\url{http://www.infoelectoral.mir.es/}}. Estos datos así creados están dispuestos de
foram tal que tan sólo hay que copiar y pegar en la macro creada por Bochsler. De esta manera
se suaviza enormemente el trabajo de cargar la macro con datos.
En la hoja excel que se crea, la última línea contiene los totales que son los que hay
que colocar en la fila 8 de la macro de Daniel Bochsler
}
\section{utilidades}{
NA
}

\examples{

Bochsler(2019,"04","D:/","D:/")
Bochsler(2016,"06","D:/","D:/")

}
