% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{competitividad}
\alias{competitividad}
\title{Indice Dimension de voto. Competitividad ( electoral, parlamentaria).}
\usage{
competitividad(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve un dataframe, con las columnas "electoral" y "parlamentario" para albergar
respectivamente el valor del índice de competitividad electoral o el parlamentario
}
\description{
Con este índice se va a medir el nivel de rivalidad electoral que existe en un
determinado sistema de partidos entre el primer y el segundo ganador de unas elecciones. Este indicador
por lo tano va a poner de manifiesto cuánta rivalidad hay entre los dos partidos que mayor número
de votos han obtenido. Para medir este fenómeno se uutiliza la aproximación o lejanía de los resultados
de las dos formaciones políticas más votadas. El calculo de este indicador se hace mediante la
formula siguiente:

\deqn{competitividad=1-(p_1-p_2)=}{competitividad=1-(p1-p2)}

Donde p1 y p2 son los porcentajes de votos ( escaños), en tanto por uno, de las dos formaciones
políticas más votadas. Un valor cercano a 1 ( p1=p2) indica fuerte competividad, mientras que un valor
de  cero (p1=1 p2=0) indicará competitividad nula.
}
\examples{

competitividad(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
