% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{concentracion}
\alias{concentracion}
\title{Indice Dimension de voto.Concentracion del voto ( electoral y parlamentario)}
\usage{
concentracion(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve un dataframe, con las columnas "electoral" y "parlamentario" para albergar
respectivamente el valor del índice de concentración de voto electoral o el parlamentario
}
\description{
Otro indicador muy importante del sistema de partidos es el de concentracón del voto,
que indica qué porcentaje de votos o escaños se llevan los dos partidos que tienen el mayor
número de votos. La fórmula utilizada para calcular este indicador es la siguiente:

\deqn{concentración=p_1+p_2}{concentración= p1+p2}

Este indicador como mucho vale 1, de tal manera que cuanto más cercano a 1 se encuentre,
mayor concentración el voto se tiene.
}
\examples{

concentracion(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
