% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Cox_Shugart}
\alias{Cox_Shugart}
\title{Indice de desproporcionalidad. de Cox_Shugart-1991 ( CS )}
\usage{
Cox_Shugart(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
El índice de desproporcionalidad de Cox_Shugart (1991) mide la desproporción en el
reparto de escaños mediante una línea de regresión entre el porcentaje de escaños y el porcentaje
de votos. Si el reparto fuera aproximadamente proporcional la pendiente de la recta de regresión
sería uno. Si se aleja de ese valor existirá desproporcionalidad en el reparto de escaños.
}
\examples{

Cox_Shugart(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
