% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Cox_Shugart_correg}
\alias{Cox_Shugart_correg}
\title{Indice de desproporcionalidad de Cox_Shugart corregido ( CS_correg )}
\usage{
Cox_Shugart_correg(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
Este índice viene a corregir el defecto que presenta el índice de Cox_Shugart, ya
que el mismo es muy sensible a la presencia de partidos políticos pequeños. Para corregir esa deficiencia
lo que se hace para calcular este indicador es no tener en cuenta los partiddos que no obtienen
representación parlamentaria. De hecho, cunado se calcula su valor lo único que se hace es
eliminar los votos y los escaños a los partidos que no tienen escaños y esos resultados se les pasa
a la función "Cox_Shugart()".
}
\examples{

Cox_Shugart_correg(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
