% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leer_Datos.R
\name{Desagregados_Mesa_Mir}
\alias{Desagregados_Mesa_Mir}
\title{Download. Datos desagregados a nivel mesa electoral obtenidos del MIR (Ministerio Interior)}
\usage{
Desagregados_Mesa_Mir(Ano, Mes, Tipo, Ruta, Borrar = T)
}
\arguments{
\item{Ano}{El año de la eleccion cuatro dígito (YYYY). Puede ser numérico o texto}

\item{Mes}{El mes de la elección. Tiene que ser como texto. (Por ejemplo "06", correspondiente a junio)}

\item{Tipo}{El tipo de fichero a descargar: "Cogreso" o "Europeas"}

\item{Ruta}{Es la ruta donde se descargarán los ficheros tipo zip del MIR}

\item{Borrar}{Es de tipo Boolean e indica si se borran o no los ficheros después de obtener el data.frame}
}
\value{
objeto de tipo tbl_df con los datos del voto a nivel de mesa electoral
}
\description{
Esta función sirve para descargar los datos en excel a nivel de mesa electoral y
despues carga un data frame con esa información.¡¡¡Observación!!!: Los campos devueltos son todos
de tipo character, por lo que si es necesario hacer operaciones, habrá que convertir
los campos necsarios a numéricos
}
\examples{

c2<-Desagregados_Mesa_Mir(2019,"04",Tipo = "Congreso", Ruta = "G:/",Borrar = T)


}
