% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indelec.R
\name{DesAgregadosIndi}
\alias{DesAgregadosIndi}
\title{utilidades. Obtencion indicadores electorales datos desagrupados.}
\usage{
DesAgregadosIndi(Ano = 0, Mes = "", RutaDescarga = "", Auto = T,
  datos_v = "", datos_d = "")
}
\arguments{
\item{Ano}{es el año del proceso electoral que se quiera tratar. Debe ser un valor
numérico de cuatro dógitos( Si se utiliza un procedimiento manual, no hace falta
ningún valor a este parámetro) .}

\item{Mes}{es el mes del proceso electoral, debe ser una cadena con dos carcateres
numéricos, asociados al mes en que se han celebrado las elecciones.( Si se utiliza
un procedimiento manual, no hace faltaningún valor a este parámetro) .}

\item{RutaDescarga}{debe ser una cadena indicativa del camino a seguir para descargar
el fichero proveniente del Ministerio del Interior. Posteriormente y de forma
automática este fichero se borra. Se hace la advertencia de que esta ruta debe
indicar un lugar donde se tenga permiso de lectura y escritura, ya que en caso contrario
no se ompletará el proceso.( Si se utiliza un procedimiento manual, no hace falta
ningún valor a este parámetro) .}

\item{Auto}{Puede tener los valores lógicos TRUE ó FALSE. Por defecto tiene
el valor TRUE, para indicar que se quiere un porceso automático. En el supuesto de
querer un proceso manual, se dará un valor FALSE a este parámetro.}

\item{datos_v}{es un data.frame  con al menos tres columnas. La primera contiene el
nombre de la unidad geográfica que agrupa los elementos de la tercera columna,
que en el caso de España, inicialmente puede ser el nombre de la Comunidad
Autónoma. La segunda columna es de carácter alfabético y contiene el código
de la unidad geográfica que aparece en la tecera columnas.En el caso de España,
lo normal será el código de provincia INE. La tercera columna será también
de tipo carácter y contiene el nombre de la unidad geográfica, en el caso de
España normalmente será el nombre de la provincia. La cuarta columna y siguientes
hará referencia  a un determinado partido político, y el nombre de la columna
se aconseja coincida con las siglas del partido político en cuestion  será de
tipo numérico y contrendrá los votos de dicho patido.  (Si se
utiliza un procedimiento automático, no hace falta ningún valor a este parámetro)}

\item{datos_d}{Es un data.frame con la misma estructura que datos_v, lo único que
en las columnas de la cuarta en adelante figurarán los diputados obtenidos por cada
partido político para cada una de las unidades territoriales contempladas en las
filas. \strong{El orden de estas columnas}, debe ser mismo que el utilizado en datos_v
para incluir los votos.(Si se
utiliza un procedimiento automático, no hace falta ningún valor a este parámetro)}
}
\value{
La salida consiste en un conjunto de listas en dos capas
\enumerate{
\item Capa de Comunidades Autónomas o unidades geográficas de agrupación. Hay una lista
para cada Comunidad Autónoma o unidad de agrupación. El orden de esta lista coincide
con el orden en que aparecen las Comunidades Autonomas o unidades de agrupación en
el data.frame de entrada. Después, elegida una Comunidad Autónoma o unidad de agrupación
se obtiene otra lista con el mismo contenido ( los mismos cuatro elementos) que el explicado para la función
}

2.- Capa de provincias o unidades desagrupadas. En este caso también se obtiene una
lista para cada provincia o unidad de desagrupación, con el identificador igual
al que figura para la provincia o unidad de agrupación en los ficheros de entrada.Finalmente
para cada provincia, se obtiene una lista con los cuatro elementos que se tenían con
la función 'AgergadosIndi()',
}
\description{
Con esta función se van a obtener los mismos resultados que con la
función 'AgregadosIndi()', pero para cada una de las regiones que figuren en
los datos facilitados. En concreto cuando se descargan datos del Ministerio del Interior
se obtendran los datos indicados, tanto para cada Comunidad Autónoma como para
cada provincia. El formato de esta salida, se indica más adelante.

\strong{NOTA IMPORTANTE:} También en esta ocasión se pueden introducir los datos de
forma automática o manual
}
\examples{
d2<-DesAgregadosIndi(2019,"04",RutaDescarga = "D:/")


}
