% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{fragmentacion_rae}
\alias{fragmentacion_rae}
\title{Indice Dmension de voto de fragmentacion ( electoral y parlamentaria) de Rae ( F )}
\usage{
fragmentacion_rae(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve un dataframe, con las columnas "electoral" y "parlamentario" para albergar
respectivamente el valor del índice de fagmentación electoral o el parlamentario
}
\description{
Este indicador pretende resumir en un valor el nivel de dispersión o concentración de
poder político, es decir y de forma resumida si se encuentra concentrado o no los escaños o votos
recibidos en una serie de partidos polítcos. Los valores del índice varían entre 0 y 1, de forma que
un valor cero indica no hay ninguna fragmentación ( todos los votos van a un sólo partido), mientras
que un valor dercano a uno indica fuerte fragmentación electoral

 La fórmula que se utiliza para su cálculo es la
siguiente:

\deqn{1-\sum_{i}q_{i}^{2}}{1-sum(qi^2)]}
for \eqn{i =  1, 2, \ldots,n }

Donde qi son las proporciones (en tanto por uno) del número de escaños o votos ( indicador electoral
o parlamentario), de los votos obtenidos por cada candidatura.
}
\examples{

fragmentacion_rae(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
