% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Gallagher}
\alias{Gallagher}
\title{Indice de los cuadrados minimos de Gallagher-1991 (Gcm)}
\usage{
Gallagher(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
El índice de desproporcionalidad electoral de los cuadrados mínimos de Gallagher, utiliza
una fórmula algo más elaborada que los índices anteriores (R, LH) para ello calcula las diferencias
al cuadrado entre los votos y escaños, los suma y el resultado lo divide entre dos y después
calcula su raiz cuadrada. Por consiguiente pondera adecuadamente las distorsiones de la
desproporcionalidad. Su fórmula matemática es la siguiente:

\deqn{Gcm=\sqrt{\frac{\sum(V_{i}-E_{i})^{2}}{2}}}{Gcm =sqrt(sum(Vi-Ei)^2*(1/2)) }
for \eqn{i =  1, 2, \ldots,n}
}
\examples{

Rae(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))
Gallagher(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))


}
