% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codigos.R
\name{get_Provincias}
\alias{get_Provincias}
\title{utilidades. Codigos de las provincias de España.}
\usage{
get_Provincias()
}
\value{
Devuelve un dataframe con dos columnas. La primera contiene el código de la
provincia y el segundo campo la denominación de esa provincia.
}
\description{
Con esta función se obtiene la denominación de todas las provincias españolas, junto
a sus códigos, que coinciden en su integridad con los que facilita el Instituto Nacional de Estadística
\href{https://www.ine.es/daco/daco42/codmun/cod_provincia.htm}{en esta página web}. Esta función no tiene parámetros.
}
\examples{
get_Provincias()



}
