% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{InjusticiaM}
\alias{InjusticiaM}
\title{Indice Dimension de voto.Injusticia matematica ( IM )}
\usage{
InjusticiaM(dates)
}
\arguments{
\item{dates}{Es un objeto de tipo data.frame que contiene la información de la siguiente
manera: La primera columna está reservada para el nombre de los partidos políticos. La
segunda columna contiene los votos obtenidos, y la tercera sirve para anotar los escaños
que obtiene ese partido político}
}
\value{
Como resultado se obtiene un objeto de tipo matrix, de dimensión n x n, con n igual
al número de partidos que se pasan en el data.frame de entrada. Los nombres de las filas
y las columnas coinciden con los nombres de los partidos pasados en el data.frame de
entrada en la primera columna.
}
\description{
Este indicador fue propuesto por Edward V.Huntington y lo que mide es el nivel
de "injusticia" que se produce en un sistema electoral, al traducir los votos conseguidos
por las formaciones políticas en escaños. En este sentido Edward V.Huntington, definió
esta injusticia matemática entre dos partidos que compiten en un proceso electoral como
la diferencia en valor absoluto entre los cocientes de escaños y votos obtenidos por
esos dos partidos políticos. Es decir para cada partido político se obtiene el cociente
entre escaños y votos, y la injusticia matemática será la diferencia en valor absoluto de
esos cocientes.

La fórmula matemática utilizada es la siguiente:

\deqn{IM_{ij}=\left|\frac{e_{i}}{v_{i}}-\frac{e_{j}}{v_{j}}\right|}{IM_ij=abs((ei/vi)-(ej/vj))}
for \eqn{i =1,  2, \ldots,n } y vi los votos del partido i y ei sus escaños
}
\examples{
a<- data.frame(par=c('A','B','C','D'),
              vot=c(200,300,100,24),sea=c(3,4,1,0), stringsAsFactors = FALSE)

InjusticiaM(a)
}
