% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilidades.R
\name{InjusticiaM_desagregada}
\alias{InjusticiaM_desagregada}
\title{utilidades. Calculo de Injusticia Matematica, desagregada

'@section utilidades}
\usage{
InjusticiaM_desagregada(Ano, Mes, Ruta)
}
\arguments{
\item{Ano}{Es el valor del año de los datos que se quieren descagar. Este dato puede ser
numérico o de tipo carácter ( mejor numérico)}

\item{Mes}{Es el mes en el que se ha realizado la elección. Obligatoriamente tiene que ser
de tipo carácter y a dos dígitos, es decir "04" es válido pero "4" no es válido}

\item{Ruta}{Se debe indicar una ruta del disco duro local donde se descarga el
fichero del Ministerior del Interior. Una vez finalizada la descarga de forma automática
el fichero descargado se borra.}
}
\value{
El resultado es una lista con 53 elementos, de manera que cada elemento es una
matriz que contiene los datos de las injusticia matemática calculada. Los 'names' de esta
lista son los códigos de provincia ( códigos INE), o bien la expresión 'Total' si son los datos
de toda España.
}
\description{
La función de este mismo paquete denominada 'InjusticiaM()', permite obtener
la injusticia matemática para el conjunto de datos que se pase. Ahora bien, si queremos
obtener la injusticia matemática para diferentes áreas geográficas se requiere un importante
esfuerzo y tiempo para preparar y ejecutar los datos. Con esta función, se agiliza enormemente
este proceso, si lo que se quiere evaluar son resultados al Congreso de los Diputados en
España, ya
que gracias a esta función, se puede extraer de forma automática la información del
Ministerio de Interior, se procesa de forma directa para cada provincia y los datos nacionales
y se devuelve el resultado en un objeto de tipo list() de R, donde cada elemento se
corresponde con el resultado obtenido de cada provincia o del Total Nacional.
}
\examples{
f<-InjusticiaM_desagregada(2019,"04","D:/")

}
