% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{INP}
\alias{INP}
\title{Indice de nacionalizacion de un partido (INP)}
\usage{
INP(dat)
}
\arguments{
\item{dat}{Vector conteniendo las proporciones ( tanto por uno ) de voto del partido en
en cada circunscripción}
}
\value{
Un número real conteniendo el valor del indicador
}
\description{
Jones y Mainwaring (2003), propusieron el cálculo de este índice basado en
el índice de concentración de la renta de Gini. Este valor va a tomar valores entre 0 y 1.
Un valor cercano a 1 indica que el partido está muy nacionalizado, y un valor cercano a
cero indicará todo lo contrario.
Este índice mide el nivel de homogeneidad de los votos que recibe un partido en todas las
circunscripciones donde se compite electoralmente. De forma muy resumida, la fórmula de
cálculo de este indicador es la siguiente.

\deqn{IPN=1- Coeficient de Gini}{IPN=1- Coeficiente de Gini}
}
\examples{

INP(c(0.15,0.22,0.24,0.26,0.27,0.27,0.27,0.32,0.33,0.34))

INP(c(0.3,0.1,0.2))

}
