% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{INSP}
\alias{INSP}
\title{Indice Nacionalizacion del Sistema de Partidos (INSP)}
\usage{
INSP(datos)
}
\arguments{
\item{datos}{Es un dataframe con la primera columna conteniendo el nombre de la circunscripción
electoral, el restos de las columnas, se corresponden con los partidos presentados y cada
columna contendrá el número de votos que cada partido ha obtenido en la circunscrición
electoral correspondiente.}
}
\value{
Un número real conteniendo el valor del indicador
}
\description{
Este índice se basa en el índice de nacionalización de cada partido. Su cálculo
se basa en el agregado de puntuaciones de nacionalización de los partidos ponderado con
el valor del peso electoral ( número de votos obtenidos).Al igual que en el índice de
Nacionalización de partidos, unos valores cercanos a 1 indicará una fuerte nacionalización
y valores cercanos a cero indicarán lo contrario. La fórmula empleada para este indicador
es la siguiente:

\deqn{INSP=sum(INPi ∗ pi)}{INSP=sum(INPi ∗ pi)}

Donde INPi es el índice de Nacionalización del partido i, y pi el tanto por uno
de votos que recibe el partido i-ésimo.
}
\examples{

a <- data.frame(prov=c("Alava","Albacete","Valladolid"),PP=c(20,30,10),PSOE=c(34,12,45),
stringsAsFactors = F)

INSP(a)


b<-data.frame(prov=c("1","2"),A=c(400,190),B=c(200,1000),stringsAsFactors = F)
INSP(b)

}
