% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{L_max}
\alias{L_max}
\title{Indice de desproporcionalidad de maxima desviacion (Lmax)}
\usage{
L_max(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
Indicador de desproporcionalidad adpatado para aquellos sistemas que utilizan la Ley
D'Hondt en el reparto de sus escaños, como es el casod de España ( salvo en las elecciones al Senado),
presenta la desventaja de que tan sólo tiene en cuenta la diferencia entre escaños y votos sólo
para la fuerza política más votada. La fórmula que se utiliza para su cálculo es la siguiente:
\deqn{Lmax=max[|V_i-E_i]}{Lmax=max[|Vi-Ei|]}
for \eqn{i =  1, 2, \ldots,n }
}
\examples{

L_max(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
