% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{L_Tukey}
\alias{L_Tukey}
\title{Indice de desproporcionalidad Linea de Tuckey (LT)}
\usage{
L_Tukey(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
Este indicador también se denomina índice de sesgo robustos, y el resultado se obtiene
mediante la utilización de la pendiente de la conocida "Linea de Tuckey" y su utilización queda recomendada
cuando exista un "pequeño" grupo de partidos que tienen un comportamiento significativamente diferente
al resto. Este  procedimiento se califica como "robusto" pues la presencia de comportamientos de
ciertos partidos diferentes al resto no altera de forma sustancial la recta obtenida, a diferencia
de lo que ocurría cuando se utilizaba en críterio de mínimos cuadrados para hacer el ajuste
de regresión.
}
\examples{

L_Tukey(c(3947,3189,1971,466,345,82),c(184,99,44,10,0,0))

}
