% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Loos_Hanby}
\alias{Loos_Hanby}
\title{Indice de desproporcionalidad de Loosemore y Hanby-1971 ("LH")}
\usage{
Loos_Hanby(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
El índice de Loosemore y Hanby ( 1971 ) pretende solventar las dificultades encontradas
con el índice de RAE. Para calcular este índice de desproporcionalidad, lo que se hace es sumar los
valores absolutos de las diferencias entre votos y escaños y dividir el resultado entre dos. La
fórmula concreta que se utiliza es la siguiente:
'\deqn{LH=\frac{\sum_{i=1}^{n}|E_{i}-V_{i}|}{2}}{LH = (1/2)*sum(|Vi-Ei|)}
for \eqn{i = 1, 2, \ldots,n}
}
\examples{
Loos_Hanby(c(1200, 30, 4000),c(10,6,8))

}
