% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapa.R
\name{mapa}
\alias{mapa}
\title{Graficos.Representacion en mapas. choropleth map}
\usage{
mapa(dat, camino, titulo, size_letra = 3, color_text = "brown",
  ver_text = T)
}
\arguments{
\item{dat}{es un data.frame con dos columnas. La primera contiene el código INE de
la provincia, y la segunda el valor de la variable que se quiera representar.
\strong{NOTA IMPORTANTE:} Se debe tener cuidado de que la primera columna no sea de
tipo factor. Se recomienda sea de tipo character, nunca pasar un factor.}

\item{camino}{Sera un valor de tipo carácter para indicar donde se encuentran los
ficheros descargados de github conteniendo los datos del mapa que se ha indicado
anteriormente. El camino debe finalizar con el caráter "/".}

\item{titulo}{Aquí se indicará el título que se desee mostrar en el mapa.}

\item{size_letra}{Es el tamaño de la letra que se utilizará para mostrar los valores
numéricos de los datos en el mapa.}

\item{color_text}{Sirve para indicar el color del texto que se utilizará para mostrar
los datos en el mapa}

\item{ver_text}{Debe ser un valor lógico. Por defecto vale TRUE, para indcar que se
muestren los datos en el mapa. Si se asigna el valor FALSE, los valores de los datos
no se mostrarán.}
}
\value{
Un objeto de tipo ggplot2 con el mapa a representar
}
\description{
Con esta función se podrá hacer mapas de tipo "choropleth map",
para hacer representaciones de los datos sobre un mapa. Inicialmente el mapa
puede ser de España, pero se podría utilizar cualquier otro de tipo "shapefile".
\strong{NOTA:} El fichero de typo shapefile que se deberá utilizar se puede bajar de la siguiente
dirección \url{https://github.com/Miguelro/Electoral/tree/master/Mapa} y se colcarán todos los
ficheros en una carpeta, que servirá para marcar el path a indicar en el
parámetro camino
}
\examples{

d <-data.frame(
cod=c(1,2,5,9,15,23,43,50),
da = c(3,5,7,10,1,3,80,4)
)
suppressWarnings(mapa(d,"F:/Elecciones/TFM/Mapa/",
                     titulo="Un mapa de ejemplo, y valores NA",ver_text = TRUE))


}
