% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{nep}
\alias{nep}
\title{Indice Dimension de voto. Numero efectivo de partidos (electoral y parlamentario ) ( N)}
\usage{
nep(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve un dataframe, con las columnas "electoral" y "parlamentario" para albergar
respectivamente el valor del índice del número efectivo de partidos  electoral o el parlamentario
}
\description{
Este indicador ( Laakso y Taagepera 1979) es complementario del indicador de fragmentación electoral, y su objetivo
es medir el número de partidos que realmente compiten en un proceso electoral. Este indicador, normalmente
oscila entre el -1,+1 del número de partidos que obtienen mas de un 10 por ciento de los votos.La
fórmula matemática que se emplea para calcular este indicador es la sigiente:

\deqn{N=\frac{1}{\sum p_{i}^{2}}}{N=1/sum(pi^2)}
for \eqn{i = 1, 2, \ldots,n } y pi los porcentajes ( en tanto por uno) de votos o escaños
del partido i
}
\examples{

nep(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
