% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{nepMolinar}
\alias{nepMolinar}
\title{Indice Dimension de voto. Indice de Numero partidos de Molinar (electoral y parlamentario ) (NP)}
\usage{
nepMolinar(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve un dataframe, con las columnas "electoral" y "parlamentario" para albergar
respectivamente el valor del índice del Numero de partidos de Molinar  electoral o el parlamentario
}
\description{
Este indicador propuesto por Molinar ( 1991), intenta evaluar el número relevante
de partidos que realmente existe en un proceso electoral. Cabe decir de este índice que presenta un
mejor comportamiento, que los índices del número de partidos e hiperfraccionamiento,tanto en la
ponderación que hace del partido ganador como en la diferencia que hay entre el primero y el
segundo partido, así como del grado de concentración de los partidos minoritarios.La formula empleada
para el cálculo de este índice es la siguiente:

\deqn{NP=1+N^{2}\sum_{i=2}p_{i}^{2}}{NP=1+N^2*sum(pi^2, for i>=2)}
for \eqn{i =  2, \ldots,n } y pi los porcentajes ( en tanto por uno) de votos o escaños
del partido i
}
\examples{

nepMolinar(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
