% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{polarizacion}
\alias{polarizacion}
\title{Dimension del voto. Indice polarizacion. ( PP )}
\usage{
polarizacion(votos, escala)
}
\arguments{
\item{votos}{Es un vector de números enteros, conteniendo los votos o escaños de todas las candidaturas.}

\item{escala}{Es un vector de número decimales entre 1 y 10  para indicar la ubicación ideológica de cada partido
de tal manera que 1 significa totalmente a la izquierda y 10 la ideología es totalmente a
la derecha.}
}
\value{
Devuelve el valor obtenido por el indicador
}
\description{
Esta dimensión es una de las más importantes en los sistemas democráticos, ya que explica
parte de los problemas de estabilidad y quiebra de las democracias.El índice que se calcula con esta
función necesita los datos en el parámetro "escala" que son difíciles de conseguir y su significado
se puede ver en el apartado de los parámetros que se expondrá más adelante. Si el indicador toma
un valor cercano a cero, significará no habrá polarización, y la misma se incrementará a medida
que el indicador tome valores más altos.
En este apartado se calcula el índice de polarización ponderado, y para su calculo se empleará
la siguiente función:

\deqn{pp=\sum_{i}p_{i}(x_{i}-\overline{x}_{p})^{2}}{pp=sum(pi*(xi-mean(x_p)))}
for \eqn{i =1,  2, \ldots,n } y pi los porcentajes ( en tanto por uno) de votos o escaños
del partido i
}
\examples{

polarizacion(c(250,346,267,876,298),c(3.2,8,7.5,6,4))

}
