% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizacion.R
\name{polarizacion}
\alias{polarizacion}
\title{Dimension del voto. Indice polarizacion. ( PP )}
\usage{
polarizacion(datos, Tipo = 1)
}
\arguments{
\item{datos}{Es un data.frame con tres columnas: la primera contiene el nombre de los
partidos. La segunda contiene el valor de la ubicación ideológica ( escala de 1 a 10),
y la tercera el número de votos obtenidos ( polarización electoral), o el número
de escaños obtenidos ( polarización parlamentaria).}

\item{Tipo}{Puede tomar los valores 1 ó 2. 1 Para indicar que se obtenga la fórmula
ponderada pero calculando las distancia a la media ponderada al cuadrado. Si se pasa el
valor de 2, se utilizará la fórmula que obtiene el valor absoluto de las diferencias
a la media ponderada}
}
\value{
Devuelve el valor numérico de este indicador
}
\description{
Esta dimensión es una de las más importantes en los sistemas democráticos, ya que explica
parte de los problemas de estabilidad y quiebra de las democracias.El índice que se calcula con esta
función necesita los datos la ubicación ideológica de los partidos que son difíciles de conseguir 
y que no es otra cosa más que una escala de valores entre 1 y 10, de tal manera que cuanto más cercano
a 1 se esté más a la izquierda estará posicionado ese partido, y el caso contrario se dará cuando
una formación política tenga una valoración en el otro extremos opuesto de  la escala. El valor 
que se obtenga para está ubicación ideológica se colocará en la segunda columna del data.frame denominado
"datos", que es el primer parámetro que se pasa a la función y cuya estructura se explica más adelante 
en el apartado dedicado a los parámetros.
Si el indicador toma un valor cercano a cero, significará no habrá polarización, y la misma se incrementará a medida
que el indicador tome valores más altos.
En este apartado se calcula el índice de polarización ponderado, y para su obtención se emplearán
las siguientes modalidades de fórmulas:

\deqn{Pp=\sum_{i=1}^{n}p_{i}(x_{i}-\overline{x}_{p})^{2}}{Pp = sum(p_j*(x_j-mean(x_p))^2)}
for \eqn{i = 1, 2, \ldots,n}

\deqn{Pp2=\sum_{i=1}^{n}p_{i}\left|x_{i}-\overline{x}_{p}\right|}{Pp2 = sum(p_j*|x_j-mean(x_p)|)}
for \eqn{i = 1, 2, \ldots,n}
}
\examples{
d <- data.frame(partidos=c("RN","PDC","PS","PPD","UDI","PRSC","otros"),
 ubicacion=c(6.36,5.31,2.73,4.13,7.04,4.00,5.33),
 c(19,20,15,21,33,7,5))
polarizacion(d,Tipo = 2)

}
