% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarizacion.R
\name{polarizacion_Dalton}
\alias{polarizacion_Dalton}
\title{Dimension del voto. Polarizacion ponderada adaptada Dalton. ( Pd )}
\usage{
polarizacion_Dalton(datos)
}
\arguments{
\item{datos}{Es un data.frame con tres columnas: la primera contiene el nombre de los
partidos. La segunda contiene el valor de la ubicación ideológica ( escala de 1 a 10),
y la tercera el porcentaje votos obtenidos respecto de todos los partidos presentados,
se tengan en cuenta o no para calcular la fórmula.}
}
\value{
Devuelve el valor numérico de este indicador
}
\description{
Con esta función se calcula el índice de polarización ponderada adaptado
de Dalton ( 2008). Este índice pondera las posiciones ideolóicas de los partidos por su
resultado electoral. Este indicador oscila entre 0 ( en el caso hipotético de que todos
los partidos ocupen la misma posición ideológica) y 10 si los partidos se encuentran en los
extremos de la posición de la escala. La fórmula de cálculo es la siguiente:
\deqn{Pd=\sqrt{\sum_{i=1}^{n}p_{i}\left[\frac{\overline{x}_{i}-\overline{x}_{p}}{4.5}\right]^{2}}}
{Pd = sqrt(sum(p_j*(x_j-mean(x_p))/4.5)^2)}
for \eqn{i = 1, 2, \ldots,n}
}
\examples{
d2<- data.frame(partidos=c("RN","PDC","PS","PPD","UDI","PRSC"),
 ubicacion=c(6.36,5.31,2.73,4.13,7.04,4.00),
 c(14.12,20.76,10.05,15.42,22.36,3.54))

polarizacion_Dalton(d2)


}
