% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regionalismo.R
\name{Regionalismo}
\alias{Regionalismo}
\title{Indice Dimension de voto. Regionalismo}
\usage{
Regionalismo(Ano = 0, Mes = "", RutaDescarga = "", Auto = TRUE,
  datos = "", PANES = "", Generate_PANES = FALSE)
}
\arguments{
\item{Ano}{es el año del proceso electoral que se quiera tratar. Debe ser un valor
numérico de cuatro dígitos( Si se utiliza un procedimiento manual, no hace falta
ningún valor a este parámetro) .}

\item{Mes}{es el mes del proceso electoral, debe ser \strong{una cadena con dos caracteres}
numéricos, asociados al mes en que se han celebrado las elecciones.( Si se utiliza
un procedimiento manual, no hace falta ningún valor a este parámetro) .}

\item{RutaDescarga}{debe ser una cadena indicativa del camino a seguir para descargar
el fichero proveniente del Ministerio del Interior. Posteriormente y de forma
automática este fichero se borra. Se hace la advertencia de que esta ruta debe
indicar un lugar donde se tenga permiso de lectura y escritura, ya que en caso contrario
no se completará el proceso.( Si se utiliza un procedimiento manual, no hace falta
ningún valor a este parámetro) .}

\item{Auto}{Puede tener los valores lógicos TRUE ó FALSE. Por defecto tiene
el valor TRUE, para indicar que se quiere un proceso automático. En el supuesto de
querer un proceso manual, se dará un valor FALSE a este parámetro.}

\item{datos}{Son los valores que serán procesados si se indica un procesamiento manual, es decir «Auto=FALSE».
Debe ser un dataframe con la siguiente estructura. La primera columna debe contener el nombre
de la unidad intermedia ( en el caso de España la comunidad autónoma), en la segunda
columna se debe incluir el nombre de las unidades inferiores ( en el caso de España, el nombre
de la provincia), y después se debe tener una columna por cada partido político que contenga
los votos obtenidos por ese partido en cada unidad inferior ( en España en cada provincia). El
valor del nombre para cada una de esta columnas debe ser las siglas que identifiquen a
cada partido político.}

\item{PANES}{Deber ser el camino completo del fichero de tipo csv que contiene para
cada partido la información sobre si dicho partido es o no de tipo regionalista.
1 será si el partido es regionalista o nacionalista y 0 en caso contrario.
Estos valores figurarán en la segunda columna de ese fichero csv.}

\item{Generate_PANES}{es un parámetro de tipo boolean, por defecto tienen el valor de FALSE. En
el caso de que sea TRUE se generará el fichero de tipo csv que se indique en el parámetro
"PANES". La primera línea de este fichero de texto será la siguiente: "PARTIDOS,PANE". que será
el nombre de las variables con las que se trabajará internamente en la función.}
}
\value{
Devuelve una lista con todos los indicadores que se obtienen con esta función.Esta
lista inicial, contiene a su vez otras dos listas.
\enumerate{
\item VRtaD. Contiene otras tres listas, cada una conteniendo un data.frame con la siguiente
información: \strong{$VRta_Provincias}: contiene los valores de indicador de VRta al nivel más
inferior ( en el caso de España, las provincias). \strong{$VRta_CCAA} contiene el indicador
Vrta a nivel de comunidad autónoma. \strong{VRtaD} Contiene el indicador VRtaD al nivel
inferior ( en el caso de España a un nivel provincial)
\item VRD. Contiene a su vez tres listas, cada una conteniendo un data.frame con la siguiente
información. \strong{$inferior_medio} Contiene el valor de VRD del nivel inferior (provincias)
respecto del nivel medio ( comunidades autónomas).  \strong{$inferior_superior} Contiene el
valor de VRD para el nivel inferior, respecto del suprior ( España). \strong{$medio_superior}
Contiene información del valor de VRD del nivel medio respecto del superior.
}
}
\description{
Con esta función calculamos los índice de regionalismo descritos en el libro
"Análisis de datos electorales" ( Pablo Oñate y Francisco A. Ocaña, pag: 48). Se remite al lector
a esta publicación para comprender el significado de los mismos. En concreto se calcularán el índice
de voto regionalista ( VRta ), el índice de voto regionalista difereneciado ( VRtaD), y el índice del
regional diferenciado ( VRD).
}
\examples{
r <- Regionalismo(Ano=2019,Mes = "11",RutaDescarga = "D:/",
 PANES = "F:/Elecciones/TFM/Datos/Regionalismo.csv")

}
