% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reparto_Escanos.R
\name{reparto_div}
\alias{reparto_div}
\title{Reparto mediante metodos de los divisores}
\usage{
reparto_div(candidaturas, votos, escanos, metodo)
}
\arguments{
\item{candidaturas}{Un vector de texto conteniendo el nombre de los partidos}

\item{votos}{Un vector de números enteros, con los votos de cada partido}

\item{escanos}{un número entero conteniendo el total de escaños a repartir}

\item{metodo}{Es el método a utilizar los valores admitidos son
c("dhondt","saint_lague","saint_lague_Mod",
"Danish","Imperiali","Hill_Huntington",
"Dean")))}
}
\value{
Devuelve un Dataframe con los nomnbre de los partidos políticos y los
Diputados asignados con el metodo elegido
}
\description{
Con esta opción se utiliza diversos métodos que tienen
Como carcaterística principal que el número de votos obtenidos por las candidaturas se
dividen por una serie de números. El nombre del método usado depende de cómo esté formada
esa serie de números. En este sentido los métodos admitidos son los siguientes:

\enumerate{
\item \bold{dhont}. Es la denominada Ley D'Hondt y es el procedimiento utilizado en España
para transformar votos en escaños en el Congreso de los Diputados. La derie de números
utilizada como conciente son los números enteros: 1,2,...n , siendo n el número de escaños
a repartir.

\item \bold{saint_lague}. Es el método de Saint Lagüe, método webster o de divisores impares.
El conjunto de divisores estárá formado por los números impares, es decir {1,3,..,2n+1}

\item \bold{saint_lague_Mod}. Es el método de Saint Lagüe modificado.Una variante del método
de Saint Lagüe, según la cual el cociente inicial es v/1.4 y a pertir de que cada lista
tenga un escaño se adopta la fórmula estándar v/(2n+1)

\item \bold{Danish}. En este método los divisores van de tres en tres unidades, es decir
está formado por los números: 1,4,7,10,13...y el n-ésimo sería 3*n-2

\item \bold{Imperiali}. Los divisores son los enteros positivos pero arranacando del
número 2 en adelante, es decir son: 2,3,4,....

\item \bold{Hill_Huntington}. En este caso los números divisores están formados por
la siguiente secuencia: sqrt(2),sqrt(6),sqrt(12),...., sqrt(n*(n+1))

\item \bold{Dean}. El conjunto de divisores está formado por los siguientes números
4/3, 12/5, 24/7,40/0,....,(2\emph{n)(n+1)/(2}n+1)

}
En todos los caso el valor de q se redondea al entero más proximo
}
\examples{

reparto_div(c("A","B","C","D","E"),c(340000,280000,160000,60000,15000),7,metodo="dhondt")
reparto_div(c("A","B","C","D"),c(340000,280000,160000,60000),7,metodo='saint_lague')
reparto_div(c("A","B","C","D"),c(340000,280000,160000,60000),7,metodo='saint_lague_Mod')
reparto_div(c("A","B","C","D"),c(340000,280000,160000,60000),7,metodo='Danish')
reparto_div(c("A","B","C","D"),c(340000,280000,160000,60000),7,metodo='Imperiali')
reparto_div(c("A","B","C","D"),c(340000,280000,160000,60000),7,metodo='Dean')


}
\seealso{
\code{\link{Restos_Mayores}} para reparto restos mayores

\url{https://en.wikipedia.org/wiki/D\%27Hondt_method} en wikipedia

Other Reparto Diputados: \code{\link{Restos_Mayores}}
}
\concept{Reparto Diputados}
