% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reparto_Escanos.R
\name{Restos_Mayores}
\alias{Restos_Mayores}
\title{Metodo restos mayores}
\usage{
Restos_Mayores(partidos, votos, escanos, metodo = "Hare")
}
\arguments{
\item{partidos}{Un vector de texto conteniendo el nombre de los partidos}

\item{votos}{Un vector de números enteros, con los votos de cada partido}

\item{escanos}{un número entero conteniendo el total de escaños a repartir}

\item{metodo}{Es el método a utilizar los valores admitidos son
c("Hare","Droop","Imperiali","Mod_Imperiali",hangenbach-bischo").
El valor por defecto es "Hare"}
}
\value{
Dataframe con los partidos políticos y los votos asignados
}
\description{
Esta función sirve para traducir una serie de votos
en escaños, siguiendo el criterio denominado "Restos mayores". De forma
resumida este método se calcula de la sigueinte forma: se divide el
número total de votos entre los escaños totales ( costo de cada escaño ).
Después se dividen los votos de cada partido entre el cociente anterior.
Se toman las partes enteras de los resultados anteriores que serán
los escaños iniciales que corresponde a cada partido. El resto de
escaños hasta completar los que hay que repartir, se asigna a los
partidos que tengan los restos más altos ( ver https://es.wikipedia.org/wiki/M%C3%A9todo_del_resto_mayor).

El costo de cada escaño es lo que diferencia a unos métodos de otros.
De esta manera si n es el total de escaños y m la suma de todos los votos,
la denominación de los métodos  es la siguiente, dependiendo del
cociente que se tome:

\enumerate{
\item \bold{Hare (Hare)}: q=m/n
\item \bold{Droop (Droop)}:q=1+(m/(1+n))
\item \bold{Imperiali (Imperiali)}: q=m/(n+2)
\item \bold{Imperiali modificado (Mod_Imperiali)}: q=m/(n+3)
\item \bold{Hangenbach Bischof (hangenbach-bischo)}: q=m/(n+1)
}

En todos los caso el valor de q se redondea al entero más proximo
}
\examples{
Restos_Mayores(c("A","B","C","D","E","F","G"),
 c(391000,311000,184000,73000,27000,12000,2000),
 21,metodo = "Imperiali")

}
\seealso{
\code{\link{reparto_div}} para reparto mediante divisores

\url{https://en.wikipedia.org/wiki/Largest_remainder_method} en wikipedia

\code{\link{reparto_div}}

Other Reparto Diputados: \code{\link{reparto_div}}
}
\concept{Reparto Diputados}
