% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Electoral_index.R
\name{Sainte_Lague}
\alias{Sainte_Lague}
\title{Indice de desproporcionalidad de Saint Lague ( SL )}
\usage{
Sainte_Lague(votes, seats)
}
\arguments{
\item{votes}{Es un vector de números enteros, conteniendo los votos de todas las candidaturas.}

\item{seats}{Es un vector de números enteros, conteniendo los escaños obtenidos por las candidaturas}
}
\value{
Devuelve el valor obtenido para el índice
}
\description{
Este índice está ideado para estudiar la desporprocionalidad de aquellos sistemas electorales
que utilizan el sistema de repartos de escaños que tiene el mismo nombre (método de Saint Lagüe).Es de
destacar que este indicador tiene en cuenta la diferencia relativa entre los escaños-votos de cada
partido, así como la del conjunto del sistema. La fórmula empleada para su cálculo es la siguiente:

\deqn{SL=\sqrt{\sum\frac{(E_{i}-V_{i})^{2}}{V_{i}}}}{SL =sqrt(sum((Vi-Ei)^2*(1/Vi)) }
for \eqn{i =  1, 2, \ldots,n vi>0}
}
\examples{

Sainte_Lague(c(3947,3189,1971,466,345,82),c(184,99,44,10,1,0))

}
