% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volatilidad.R
\name{volatilidad}
\alias{volatilidad}
\title{Indice Dimension de voto. Volatilidad}
\usage{
volatilidad(dat1, dat2, enlace)
}
\arguments{
\item{dat1}{Es un dataframe con dos columnas. La primera columna contiene las siglas
del partido político y la segunda los votos o escaños conseguidos del primer
periodo de tiempo, dependiendo
se quiera   calcular la volatilidad electoral o parlamentaria respectivamente.}

\item{dat2}{Es un dataframe con dos columnas. La primera columna contiene las siglas
del partido político y la segunda los votos o escaños conseguidos del segundo
periodo de tiempo, dependiendo
se quiera   calcular la volatilidad electoral o parlamentaria respectivamente.}

\item{enlace}{es un data.frame que sirve para enlazar los partidos, coaliciones o
agrupaciones que se quieran comparar ente los dos periodos objeto de estudio.Este
data.frame contiene un total de 22 columnas. La primera columna contiene la denominación
pertinente. Las 10 columnas siguientes (denominadas p1_i para i=1,2,...,10) sirven para
indicar el número de la/las fila/filas del/los partidos del primer periodo electoral que se tengan
que agrupar. Las 10 columnas siguientes (denominadas p2_i para i=1,2,...,10) sirven
para indicar  el número de la/las fila/filas del/los partidos del segundo periodo electoral que se tengan
que agrupar. Cuando no se necesiten las 10 columnas, las sobrantes se rellenan a ceros.
Los datos que aparecen en cada fila sirven para lo siguente: Para el primer bloque
de 10 columnas se suman los votos/escaños de los partidos que aparecen en ese bloque.
Para el segundo boque de 10 columnas, igualmente se suman los votos/escaños que aparecen
en ese segundo bloque. Estos dos valores serán los que posteriormente sirvan para
hacer la comparación.
La última columna de este data.frame ( que tiene por denominación "bloque"), contendrá
los valores "D" ó "I" indicativos de "Derecha","Izquierda"}
}
\value{
Devuelve una lista con dos objetos. El primero es la volatilidad total, y el
segundo la volatilidad entre-bloques.
}
\description{
Con esta función se calcula la volatilidad total de Pederson (1983:31 y 32),
y la volatilidad entre bloques. Las fórmulas para calcular estos valores se pueden encontrar
en el libro de Oñate y Ocaña titulado "Análisis de datos electorales", página 45.
Con esta dimensión lo que se hace es comparar el commportamiento del electorado en
dos elecciones distintas, para ver el traspase de votos, bientre entre bloques, o
entre patidos.
}
